generate completions
