+++ Cargo.toml
@@ -25,6 +25,10 @@ zip         = { version =  "0.5.13", default-features = false, features = ["defl
 flate2      = { version = "1.0.22", default-features = false, features = ["zlib"] }
 zstd        = { version = "0.9.0", default-features = false, features = ["thin"] }
 
+[build-dependencies]
+clap = "=3.0.0-beta.5"
+clap_generate = "=3.0.0-beta.5"
+
 [dev-dependencies]
 tempfile = "3.2.0"
 infer = "0.5.0"
