+++ src/archive/zip.rs
@@ -11,7 +11,7 @@ use zip::{self, read::ZipFile, ZipArchive};
 
 use crate::{
     info,
-    utils::{self, dir_is_empty,strip_cur_dir, Bytes},
+    utils::{self, dir_is_empty, strip_cur_dir, Bytes},
 };
 
 use self::utf8::get_invalid_utf8_paths;
