+++ src/lib.rs
@@ -5,18 +5,20 @@
 //! 2. It's required by some integration tests at tests/ folder.
 
 // Public modules
-pub mod cli;
+pub mod archive;
 pub mod commands;
 
 // Private modules
-pub mod archive;
+mod cli;
 mod dialogs;
 mod error;
 mod extension;
 mod macros;
+mod opts;
 mod utils;
 
 pub use error::{Error, Result};
+pub use opts::{Opts, Subcommand};
 
 /// The status code ouch has when an error is encountered
 pub const EXIT_FAILURE: i32 = libc::EXIT_FAILURE;
